simplify tests scaffolding
