+++ packages/sdks/e2e/react-native/package.json
@@ -10,7 +10,6 @@
     "web": "expo start --web",
     "build": "echo 'no need to build react-native.'",
     "dev": "yarn run web",
-    "e2e": "SDK=reactNative ../../node_modules/.bin/playwright test",
     "serve": "yarn run web"
   },
   "dependencies": {
