--- packages/sdks/e2e/react-native/playwright.config.ts
@@ -1,3 +1,0 @@
-import { configFor } from '@builder.io/sdks-e2e-tests/shared.config';
-
-export default configFor('e2e-react-native', 19006);
