+++ packages/sdks/e2e/react/package.json
@@ -6,7 +6,6 @@
   "scripts": {
     "build": "tsc --noEmit && vite build",
     "dev": "tsc --noEmit && vite",
-    "e2e": "SDK=react ../../node_modules/.bin/playwright test",
     "serve": "vite preview"
   },
   "dependencies": {
