+++ packages/sdks/e2e/rsc/package.json
@@ -5,8 +5,7 @@
   "scripts": {
     "dev": "shopify hydrogen dev",
     "build": "shopify hydrogen build",
-    "serve": "shopify hydrogen preview",
-    "e2e": "SDK=rsc ../../node_modules/.bin/playwright test"
+    "serve": "shopify hydrogen preview"
   },
   "devDependencies": {
     "@shopify/cli": "3.18.0",
