+++ packages/sdks/e2e/rsc/vite.config.ts
@@ -9,4 +9,5 @@ export default defineConfig({
   optimizeDeps: {
     include: ['@headlessui/react', 'clsx', 'react-use', 'typographic-base'],
   },
+  server: { port: 5000 },
 });
