+++ packages/sdks/e2e/solidjs/package.json
@@ -7,8 +7,7 @@
     "start": "vite",
     "dev": "vite",
     "build": "vite build",
-    "serve": "vite preview",
-    "e2e": "SDK=solid ../../node_modules/.bin/playwright test"
+    "serve": "vite preview"
   },
   "license": "MIT",
   "devDependencies": {
