--- packages/sdks/e2e/solidjs/playwright.config.ts
@@ -1,3 +1,0 @@
-import { configFor } from '@builder.io/sdks-e2e-tests/shared.config';
-
-export default configFor('e2e-solidjs', 7508);
