+++ packages/sdks/e2e/svelte/package.json
@@ -7,7 +7,6 @@
     "build": "yarn run check && vite build",
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "dev": "vite",
-    "e2e": "SDK=svelte ../../node_modules/.bin/playwright test",
     "serve": "vite preview"
   },
   "dependencies": {
