+++ packages/sdks/e2e/vue-3/package.json
@@ -6,7 +6,6 @@
   "scripts": {
     "build": "vue-tsc --noEmit && vite build",
     "dev": "vue-tsc --noEmit && vite",
-    "e2e": "SDK=vue ../../node_modules/.bin/playwright test",
     "serve": "vite preview"
   },
   "dependencies": {
