+++ packages/sdks/package.json
@@ -35,12 +35,12 @@
     "e2e:build:qwik": "yarn workspaces foreach --verbose --include \"@builder.io/e2e-qwik*\" run build",
     "e2e:build:react": "yarn workspaces foreach --verbose --include \"@builder.io/e2e?(-nextjs)-react\" run build",
     "e2e:build:vue": "yarn workspaces foreach --verbose --include \"@builder.io/e2e-vue*\" run build",
-    "e2e:run:svelte": "yarn workspaces foreach --verbose --include \"@builder.io/e2e-svelte*\" run e2e",
-    "e2e:run:react-native": "yarn workspaces foreach --verbose --include \"@builder.io/e2e-react-native*\" run e2e",
-    "e2e:run:solid": "yarn workspaces foreach --verbose --include \"@builder.io/e2e-solid*\" run e2e",
-    "e2e:run:qwik": "yarn workspaces foreach --verbose --include \"@builder.io/e2e-qwik*\" run e2e",
-    "e2e:run:react": "yarn workspaces foreach --verbose --include \"@builder.io/e2e?(-nextjs)-react\" run e2e",
-    "e2e:run:vue": "yarn workspaces foreach --verbose --include \"@builder.io/e2e-vue*\" run e2e",
+    "e2e:run:svelte": "SDK=svelte yarn @builder.io/sdks-e2e-tests run test",
+    "e2e:run:react-native": "SDK=react yarn @builder.io/sdks-e2e-tests run test",
+    "e2e:run:solid": "SDK=solid yarn @builder.io/sdks-e2e-tests run test",
+    "e2e:run:qwik": "SDK=qwik yarn @builder.io/sdks-e2e-tests run test",
+    "e2e:run:react": "SDK=react yarn @builder.io/sdks-e2e-tests run test",
+    "e2e:run:vue": "SDK=vue yarn @builder.io/sdks-e2e-tests run test",
     "e2e:build:all": "yarn workspaces foreach --exclude @builder.io/e2e-rsc --verbose run build",
     "e2e:run:all": "yarn workspaces foreach --exclude @builder.io/e2e-rsc --verbose run e2e",
     "release-sdk": "bash ./scripts/release-sdk.sh",
